feat(generator): split scaffold functionality out
